"#43 Topmenu layout implemented"
"Layout of topmenu implemented. I created a new class named MenuButton
to create a general style for all the buttons in topmenu. Added four of
them into the flow layout of topmenu and changed the background color of
topmenu into a dark color. None of these buttons is starting a function
this commit was just for styling and layouting."