"#1#17 Chessboard-Fields and Chessboard-Coordinates implemented"
"When you open the program you will see a Chessboard now with 64 fields
of two different colors and a border with typical chess coordinates on
it. The board is part of a JPanel which uses a GridBagLayout. Each field
get an ID, playing fields starts with 0 (top left field) and ends with
63 (bottom right field). All border fields have the id 100. Row/Column
values starts at -1/-1 (top left corner field) and ends at 8/8 (bottom
right corner field). So playing field with id = 0 is at row = 0 and
column = 0. Another example for the border field c: ID = 100, Row = -1,
Column = 2."