"#26#27 Long and short rochade implemented"
"King is able now to use a long or short rochade if the King and the
Rook wasn't moved before and there is a free route between them. To use
the short or long rochade the user has to click on the field where the
King will find place after the rochade movement. For the short rochade
this is +2 fields from King position and for the long rochade this is -2
fields from King position."